tests/dup2.c: print syscall() result using %ld format

+++ tests/dup2.c
@@ -12,10 +12,10 @@ main(void)
 {
 	const long int fd_old = (long int) 0xdeadbeefffffffff;
 	const long int fd_new = (long int) 0xdeadbeeffffffffe;
-	int rc = syscall(__NR_dup2, fd_old, fd_new);
-	printf("dup2(%d, %d) = %d %s (%m)\n",
-	       (int) fd_old, (int) fd_new, rc,
-	       errno2name());
+
+	long rc = syscall(__NR_dup2, fd_old, fd_new);
+	printf("dup2(%d, %d) = %ld %s (%m)\n",
+	       (int) fd_old, (int) fd_new, rc, errno2name());
 
 	puts("+++ exited with 0 +++");
 	return 0;
